Refactor: #44 상태 모달 생성/수정 폼 공유를 위한 리팩토링 (#46)
* Fix: #44 mock 데이터 에러 해결

* Refactor: #44 상태 모달 생성/수정 폼 공유를 위한 리팩토링

* Chore: #33 기능명세서 변경으로 인한 색상 추가 기능 제거

* Formatting: #44 컴포넌트 Props 이름 변경
