+++ src/constants/projectStatus.ts
@@ -1,0 +1,10 @@
+export const PROJECT_STATUS_COLORS = Object.freeze({
+  RED: '#c83c00',
+  YELLOW: '#dab700',
+  GREEN: '#237700',
+  BLUE: '#00c2ff',
+  ORANGE: '#ff7a00',
+  PURPLE: '#db00ff',
+  PINK: '#ff0099',
+  YELLO_GREEN: '#8fff00',
+});
